refactor(recommender): consolidate CLI into root main.py
